# [chore][pkg/ottl] Use faster json parsing library