--- a/pkg/ottl/ottlfuncs/func_parse_json_test.go
+++ b/pkg/ottl/ottlfuncs/func_parse_json_test.go
@@ -201,3 +201,63 @@ func Test_ParseJSON_Error(t *testing.T) {
 	_, err := exprFunc(context.Background(), nil)
 	assert.Error(t, err)
 }
+
+const benchData = `{
+	"_id": "667cb0db02f4dfc7648b0f6b",
+	"index": 0,
+	"guid": "2e419732-8214-4e36-a158-d3ced0217ab6",
+	"isActive": true,
+	"balance": "$1,105.05",
+	"picture": "http://example.com/1",
+	"age": 22,
+	"eyeColor": "blue",
+	"name": "Vincent Knox",
+	"gender": "male",
+	"company": "ANIVET",
+	"email": "vincentknox@anivet.com",
+	"phone": "+1 (914) 599-2454",
+	"address": "483 Gerritsen Avenue, Succasunna, Massachusetts, 7803",
+	"about": "Elit aliqua qui amet duis esse eiusmod cillum proident quis amet elit tempor dolor exercitation. Eu ut tempor exercitation excepteur est. Lorem ad elit sit reprehenderit quis ad sunt laborum amet veniam commodo sit sunt aliqua. Sint incididunt eu ut est magna amet mollit qui deserunt nostrud labore ad. Nostrud officia proident occaecat et irure ut quis culpa mollit veniam. Laboris labore ea reprehenderit veniam mollit enim et proident ipsum id. In qui sit officia laborum.\r\nIn ad consectetur duis ad nisi proident. Non in officia do mollit amet sint voluptate minim nostrud voluptate elit. Veniam Lorem cillum fugiat adipisicing qui ea commodo irure tempor ipsum pariatur sit voluptate. Eiusmod cillum occaecat excepteur cillum aliquip laboris velit aute proident amet.\r\nIpsum sunt eiusmod do ut voluptate sit anim. Consequat nisi nisi consequat amet excepteur ea ad incididunt pariatur veniam exercitation eu ex in. Incididunt sint tempor pariatur Lorem do. Occaecat laborum ad ad id enim dolor deserunt ipsum amet Lorem Lorem. Cillum veniam labore eu do duis.\r\nCillum dolor eiusmod sit amet commodo voluptate pariatur ex irure eu culpa sunt. Incididunt non exercitation est pariatur est. Incididunt mollit Lorem velit ullamco excepteur esse quis id magna et ullamco labore. Laboris consequat tempor est ea amet enim et nisi amet officia dolore magna veniam. Nostrud officia consectetur ea culpa laborum et ut Lorem laboris.\r\nDeserunt labore ullamco dolor exercitation laboris consectetur nulla cupidatat duis. Occaecat quis velit deserunt culpa nostrud eiusmod elit fugiat nulla duis deserunt Lorem do. Proident anim proident aute amet pariatur et do irure. Ad magna qui elit consequat sit exercitation sit. Magna adipisicing id esse aliqua officia magna. Et veniam aliqua minim reprehenderit in culpa. Adipisicing quis eu do Lorem cupidatat consequat ad aute quis.\r\nIn aliquip ea laborum esse dolor reprehenderit qui sit culpa occaecat. Consectetur Lorem dolore adipisicing amet incididunt. Dolor veniam Lorem nulla ex. Eiusmod amet tempor sit eiusmod do reprehenderit proident sit commodo elit cupidatat.\r\nNulla nulla consequat cillum mollit tempor eiusmod irure deserunt amet et voluptate. Fugiat et veniam culpa eiusmod minim ex pariatur. Eiusmod adipisicing pariatur pariatur adipisicing in consequat cillum ut qui veniam amet incididunt ullamco anim.\r\nDolor nulla laborum tempor adipisicing qui id. Exercitation labore aliqua ut laborum velit cupidatat officia. Est qui dolor sint laboris aliqua ea nulla culpa.\r\nAute reprehenderit nulla elit nisi reprehenderit pariatur officia veniam dolore ea occaecat nostrud sunt fugiat. Cillum consequat labore nostrud veniam nisi ea proident est officia incididunt adipisicing qui sint nisi. Ad enim reprehenderit minim labore minim irure dolor. Voluptate commodo dolor excepteur est tempor dolor sunt esse fugiat ea eu et.\r\nIpsum sit velit deserunt aliqua eu labore ad esse eu. Duis eiusmod non exercitation consequat nulla. Enim elit consectetur pariatur sunt labore sunt dolore non do. Sint consequat aliqua tempor consectetur veniam minim. Veniam eu aute occaecat consectetur dolore ullamco dolore officia.\r\n",
+	"registered": "2023-06-08T12:29:06 +07:00",
+	"latitude": -59.802339,
+	"longitude": -160.473187,
+	"tags": [
+		"pariatur",
+		"anim",
+		"id",
+		"duis",
+		"fugiat",
+		"qui",
+		"veniam"
+	],
+	"friends": [
+		{
+			"id": 0,
+			"name": "Hester Bruce"
+		},
+		{
+			"id": 1,
+			"name": "Laurel Mcknight"
+		},
+		{
+			"id": 2,
+			"name": "Wynn Moses"
+		}
+	],
+	"greeting": "Hello, Vincent Knox! You have 1 unread messages.",
+	"favoriteFruit": "apple"
+}`
+
+func BenchmarkParseJSON(b *testing.B) {
+	ctx := context.Background()
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_, err := parseJSON(ottl.StandardStringGetter[any]{
+			Getter: func(_ context.Context, _ any) (any, error) {
+				return benchData, nil
+			},
+		})(ctx, nil)
+		require.NoError(b, err)
+	}
+}