This follows https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/33785 by switching to a faster json library.