@atoulme 2024-09-11
Any benchmarks?