@djaglowski 2024-09-11
@BinaryFissionGames pointed out to me offline that the benchmark was just getting the function but not executing it.

Now I see a nice improvement:

```
benchstat pkg/ottl/old.txt pkg/ottl/new.txt                                                                                              ✭ ✱
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/ottlfuncs
             │ pkg/ottl/old.txt │          pkg/ottl/new.txt           │
             │      sec/op      │   sec/op     vs base                │
ParseJSON-10        278.0µ ± 2%   246.3µ ± 3%  -11.38% (p=0.000 n=10)

             │ pkg/ottl/old.txt │           pkg/ottl/new.txt           │
             │       B/op       │     B/op      vs base                │
ParseJSON-10       26.84Ki ± 0%   10.89Ki ± 0%  -59.43% (p=0.000 n=10)

             │ pkg/ottl/old.txt │          pkg/ottl/new.txt          │
             │    allocs/op     │ allocs/op   vs base                │
ParseJSON-10         206.0 ± 0%   169.0 ± 1%  -17.96% (p=0.000 n=10)
```